requirements.txt
requests
opencv-python
google-genai
pdoc